refactor(agents): split into module directory with SRP extraction
Convert agents.rs to agents/ module directory and extract:
- types.rs: AgentStatus, SessionInfo, HookStatus, RulesInfo
- util.rs: tilde_path, count_files_recursive, dir_size_human, etc.
- formatting.rs: print_text, print_json, print_help
- detection.rs: detect_all_agents, detect_* functions, settings helpers

mod.rs retains only run(), command(), module declarations, and re-exports.
Cross-module imports updated to use crate::cmd:: paths.
